requirements.txt
discord.py==2.3.2
Flask==2.1.2
requests==2.26.0